requirements.txt
requests
pyyaml
pycryptodome
tqdm